Engine gets its own Module
